Create step6
Modify the event handler for mouse clicks to flip cards based on the location of the mouse click. If the player clicked on the ith card, you can change the value of
exposed[i] from False to True
True. If the card is already exposed, you should ignore the mouseclick. At this point, the basic infrastructure for Memory is done.